Menu parsing updates for plugins.
FIFO menus are the only plugin that takes advantage of this.
Example:
<menu id="root" label="Openbox 3">
  <menu id="fonk" label="fonk" plugin="fifo_menu">
  </menu>
</menu>

This creates a FIFO ~/.openbox/fifo_menu/fonk to which you can send
menus to. The menus sent to it must be like
<fifo>
  <item>
   etc...
</fifo>

I think. If my memory serves me right.

It is all hideous, but I just wanted to experiment and see if it was
possible.