add --reconfigure option to openbox.
this option will open the display, find the current openbox's PID from the root window, and send a SIGUSR2 to it, telling it to reconfigure. then it exits.

no longer printing help if you pass an invalid argument to openbox

since that is the case, session_startup doesnt need to be modifying the args anymore