fix for weird group transient crap.
  2 group transients
+ 1 direct transient of one of the above
--
the other group transient would be a child of the direct transient.

it's friggin annoying to look for this. i had to destroy my super clever but
impossible to maintain client_update_transient_tree in order to do it, and add
redundant execution, where children update their transient trees when a parent
does.