make fullscreen windows more xinerama and multi-screen aware
instead of calling it every time a window loses focus, it is only called when a window gains focus.  then, check fullscreen layered windows, if they should be moved to a lower layer.
when moving a window between monitors, also check its layer and that of any fullscreen windows.
let a window stay in the fullscreen layer even when it is not focused, if it is on a non-visible desktop, or if it is on a different monitor from the focused window, or if nothing else is focused