Fixing bug from commit 041d17373e04
Pressing a button and leave/enter would cause the button to show hover mode, not pressed mode.  Change the behaviour back to how it used to be for pressing (the button stays pressed when you move outside of its box) and make it work correctly, as commit 041d17373e04 also did for menus.

Reverting this behaviour because it seems impossible to do the enter/leave stuff correctly for the close button on maximized windows.  Leaving the titlebar contexts doesn't give us an Enter event to go along with it, so even if we check all motion events, the button will flash unpressed when leaving the topright contexts.