Make it possible for an action name to choose whether it is interactive or not based on its options.
This way we can use the same name with options for an interactive action and a
non-interactive action.

Shorten the names of the ObActionsInteractive* functions to ObActionsI*

Add a ObActionsIPreFunc that is called for interactive actions
before the interactivity (key/mouse grab) is started.

Add a ObActionsIPostFunc that is called for interactive actions
after the interactiviti (key/mouse grab) has ended.