Specify the GLX_TEXTURE_TARGET_EXT when creating a glXPixmap
This is a workaround for non-spec drivers.  See issue #4 here: http://www.opengl.org/registry/specs/EXT/texture_from_pixmap.txt

A default should be chosen, but it is not.  We assume that GL_ARB_texture_non_power_of_two is present for now, and just always ask for a GL_TEXTURE_2D texture