don't free lists in the linkbase hash table prematurely.
the values (GSList objects) are freed by the hash table's free function if a
new value is inserted.  if we change the list and reinsert it, then it would
cause the list we're inserting to get freed (partially).  not good.  so instead
we manually free them when we remove them/destroy the hash table.