Show the list of categories in the apps menus as submenus.
The categories are not based on what you actually have yet but will be soon.

Removed the "main category" notion from the links themselves, they just
publish a list of categories instead.  Moved this notion out to the Apps menu
itself.  This should make it easy to possibly to customize your set of
visible categories if we should like to do that sometime.