window.inactive.label.text.color should default to white.
Currently it defaults to itself, which is NULL, and causes a crash.

The spec http://openbox.org/wiki/Help:Themes#window.inactive.label.text.color
claims that the default is white.

In 98b02c6b60bbde2a5db026b3ee3e6e6dc44d1a92 we had this property being read
twice, once with the NULL backup and once with white. It seems that we kept
the wrong one, so just switching this to white now.