Use PANGO_UNDERLINE_SINGLE for rendering shortcuts
The pango docs say to use _LOW for accelerators, but that causes the line
to be drawn beneath the ink extents which for us may be outside the area
that gets displayed and we end up with no line at all. We also don't want
to extend every menu entry with enough space to fit this possible line.