Implemented 'stop_hidden' configuration option for applications.
This option lets user to stop application processes automatically when all
the application windows are hidden from the user. A window is considered hidden
when it's iconified or shaded or does not belong to the current desktop.

This option in intended for minimization of CPU usage and power consumption.
Some modern applications (like Web browsers) tend to consume significant amount
of CPU resources even when hidden.

There are three possible values of the option:
   'no'      - do not stop application when it hides (default value)
   'process' - stop just the process the hidden window belongs to
   'group'   - stop whole process group the application belongs to; some complex
               applications (like web browsers) emerge multiple child processed,
               so in order to stop such applications it's not enough just to
               stop one process.

Example:

  <application class="Firefox">
    <stop_hidden>group</stop_hidden>
  </application>

Please note that Firefox does not create a new group for its child
processes. You need to run it via `setsid firefox` in order to have a separate
group for all Firefox processes.